Move WriteData to ts
