+++ ui/src/shared/components/DatabaseDropdown.js
@@ -30,7 +30,9 @@ class DatabaseDropdown extends Component {
 
     return (
       <Dropdown
-        items={databases.map(text => ({text}))}
+        items={databases.map(text => ({
+          text,
+        }))}
         selected={database || 'Loading...'}
         onChoose={onSelectDatabase}
         onClick={onStartEdit ? onStartEdit : null}
@@ -50,11 +52,15 @@ class DatabaseDropdown extends Component {
 
       const nonSystemDatabases = databases.filter(name => name !== '_internal')
 
-      this.setState({databases: nonSystemDatabases})
+      this.setState({
+        databases: nonSystemDatabases,
+      })
       const selectedDatabaseText = nonSystemDatabases.includes(database)
         ? database
         : nonSystemDatabases[0] || 'No databases'
-      onSelectDatabase({text: selectedDatabaseText})
+      onSelectDatabase({
+        text: selectedDatabaseText,
+      })
     } catch (error) {
       console.error(error)
       onErrorThrown(error)
